refactor: clean up unused code and reorganize component imports
