+++ discord/src/components/ErrorBoundary.tsx
